Only prevent default for `#` if popup shown
We were preventing the default action on keydown in EditInputGroup
if the key was enter or #. However, there were some additional checks
before the reference popup would be shown, like that the previous
character is whitespace. The event should only have the default action
prevented if the popup will be shown. Otherwise, it is impossible to
type #.

Signed-off-by: Marcel Robitaille <mail@marcelrobitaille.me>
Signed-off-by: Christian Wolf <github@christianwolf.email>