Fix # appearing in popup not input when linking
The previous fix of switching to keydownwas not viable. The # would
appear in the popup instead of in the main input. This would show no
results in the popup since none have #, and would insert the wrong thing
into the main input.

Fix this by reverting the linking event handler back to keyup. This has
fewer unintended consequences. I also think that the flow is cleaner now
that enter and # are handled in two separate functions.

Signed-off-by: Marcel Robitaille <mail@marcelrobitaille.me>