Revert "Removed obsolete parameters from OpenAPI specs"
This reverts commit d10eb6d0ada76d7b905b88f986d6dfeee4718e74.

The @type annotation should be still part of the OpenAPI description. Although the returned
JSON objects are considered compatible with the schema.org standard, this does not mean, we
should not make the implications explicit. Also, this makes reading more clean and avoid
misunderstandings regarding if the @type is to be present or not.

Signed-off-by: Christian Wolf <github@christianwolf.email>