Fix recipe import from some WordPress sites
I want to import recipes from a site that uses WordPress Recipe Maker

At least on this site, the schema is structured like so:

{
  "@schema": "https://schema.org/",
  "@graph": [
    { /* blah */ },
    { /* blah */ },
    { "@type": "Recipe", "otherStuff": True }
  ]
}

Notably missing is the @schema on the Recipe object. It's intended
to be inherited from the parent, but HttpJsonLdParser didn't support
this kind of inheritance.

Well, now it does! =)

Signed-off-by: Nathaniel <I@nathaniel.land>