Update ref in README.md
In the FAQ at the bottom of README.md is a Q&A about issues with
the recipe parser, suggesting that viewers can work on the mehod
that does the actual parsing. The link to the method is outdated.

The place I started when searching for my issue with the parser
was in `RecipeExtractionService.php` in the `parse()` method, so
in this commit I update the outdated reference to point there.

Signed-off-by: Nathaniel <I@nathaniel.land>